More precise list of uniqueness constraints to check in nodeApplyChanges
Previously too many uniqueness constraints were checked which resulted
in unnecessary index entry locks being acquired.

Due to the way the SchemaCache "get related" methods are currently
implemented, somewhat complex in nature, this new use case
of getting related uniqueness constraints for both changed labels
and changed properties isn't well catered for and is not a perfect fit.
And the selection isn't as precise as it could be, given the data
that is available at the time of asking. These SchemaCache methods
should be revisited to do this better.