Fix history bug in cypher shell
The first time you connected to cypher-shell (without providing username and password via -u and -p),
we prompt the user for username and password and that means we will load the history before the history
file was set. This leads to that the history file wasn't loaded until you explicitly call `:history`. This
PR adds an explicit `load` also when we set the history file.

NOTE: I haven't been able to figure out how to test this, we need to be able to test native signals such as
arrow keys etc.