Introduce new store format versions behind setting
This introduces a new version of standard, aligned and high_limit
formats, but will not start to use them. They are only used with
the unsupported.dbms.include_dev_record_format_versions setting enabled.
That setting will also override the version (not format family) in
dbms.record_format to be able to get the latest version of all the families.