Change KeyPartitioning to return sorted edges
Before KeyPartitioning::partition accepted an implicitly sorted
collection of key candidates, and returned a list of pairs of keys
corresponding to the edges of each partition range:
        fromInclusive, toExclusive : for each partition

Though it duplicated KEY references, it was conceptually useful to keep
the edges together; which could then could be transformed into a Seeker
(or a Seeker.WithContext) more easily.

Whilst sketching out a solution to adapt the PartitionedScan interface,
to better meet the requirements of GDS; it would be benefitial to have
a sorted list of edges, rather than pairs of edges. This also reduces
allocations.

Seekers/Seeker.WithContext can be constructed by looping over the
index of returned list of partition edges, getting the current and next.

        for ( int i = 0; i < partitionEdges.size() - 1; i++ )
        {
                KEY fromInclusive = partitionEdges.get( i );
                KEY toExclusive = partitionEdges.get( i + 1 );
                try ( Seeker<KEY,VALUE> seeker = seekerFactor.seek( fromInclusive, toExclusive, cursorContext ) )
                {
                        // ...

Whilst here, the class was tidied, and used some newer JDK features.

Tests needed fixing to handle new interface.

GBPTree::partitionedSeekInternal was modified to deal with the change of
interface, yet keep it's own; which will be updated in the next commit.