Propagate partition edges up to Partitioned{TokenScan,ValueSeek}
Changing the interface of GBPTree::partitionedSeek to return a sorted
list of partition edges, rather than a collection of Seeker.WithContexts

The implementation of PartitionedTokenScan and PartitionedValueSeek,
required a few changes. Rather than pulling from an iterator via a
synchronised method; we store the list, and use an atomic int to track
the partition's fromInclusive edge.

Having access to the partition edges at this level, rather than
Seeker.WithContexts, allows us to mutate them; specifically realign
them, depending on the index, if needed. Which should satisfy GDS
requirements.

Having them as a list of edges, rather than a list of pairs of edges,
optimises the number of function calls to mutate the edges.

Seeker.WithContext was added to handle constructing the Seeker on the
correct execution thread; as we have the edges, we can directly
construct the Seeker on the correct thread, and thus we can remove
Seeker.WithContext in this refactor.