Prevent logging appenders and log contexts memory leak
Before this PR lots of default console appenders were never closed. As a result of that map in AbstractManager was constantly growing and eventually causing OOM.
In addition to appenders, there were a couple of contexts that were never closed.
This PR start closing leaked log4j contexts and introduce a generic neo4j config that
does not create a default console logger.