Preallocate checkpoint log file
Before this PR checkpoint, log file was not preallocated and potentially could face OOD
when tried to append to the tail of log file as part of checkpoint process.
With proposed changes file is preallocated on platforms that support that and append will happen into already reserved space.
While this PR does not remove that scenario completely (preallocation can be off or not available) it makes it much less possible in the common scenario.