Use switch in IndexCapability implementations
Due to the size of the arrays, the stream overhead is not negligible.

Although the EnumSets are inlined by the optimiser, and the comparisons
are cheap; the switch will compile to a LOOKUPSWITCH / TABLESWITCH,
which should potentially be marginally more efficient; and can be done
in a single pass.

SupportPartitionedScanTestSuite confirms the behaviour is consistent.

Switched a few cases to using the new switch expressions, where
benefitial for readability.