Decrease the number of modules that neo4j requires to be open.
As a result of this change, neo4j will require only 3 open modules to operate: java.nio, java.io and sun.nio.ch.
That's applies to the default set of tests and for neo4j distributions.

As part of this PR tests that are using jamm to verify memory estimations were isolated and, additional module exposures were added only
for that module since it heavily relies on reflection and accessing all the internal details.
Please DO NOT expand usage of jamm outside of that module and add any test that requires memory measurement to that specific module.