Schema store migration to 5.0 - not enabled yet
Introduces a schema store migration that makes sure that
the node label index is persisted in the store and has
properties associated with it. Depending on the kernel version
of the old store, it could either be injected (before 4_3_D3) or have a schema
record with no properties (from 4_3_D3 and not dropped).

The migration will be run when migrating to the 5.0 store format versions
that are not used yet. To test it out the
unsupported.dbms.include_dev_record_format_versions setting can be enabled.