Create wrapper for BTREE RANGE of PointValue to BOUNDING_BOX
    (This commit is written with the intent it will be reverted.)

Though most PropertyIndexQuery::range calls for PointValue have been
updated to PropertyIndexQuery::boundingBox; Cypher will be calling
PropertyIndexQuery::range for PointValue types.

To play it safe, this wrapper class should keep the same behaviour for
geometric ranges on BTREE indexes, to be bounding box queries.

It can be removed with BTREE, or when changes to Cypher allow for
calling PropertyIndexQuery::boundingBox where appropriate.