Rotate checkpoint log file on the recovery
Allow checkpoint to survive the corruption of checkpoint records.
In a situation when we fail to read checkpoint records - use the last readable to perform recovery, if possible. If that's not possible we fail as before.
Also, as soon as we detect that after the last checkpoint there are some unreadable bytes - we will perform checkpoint log file rotation and will write a new checkpoint to the new log file.
All of that to prevent any future readers from reading that unreadable data as much as possible.