Fix flaky ClusterReconciledTransactionTrackerIT
The test case assumed nothing was written to the system db between startup and assertions. With the addition of the
TopologyGraphUpdater this assumption is no longer valid as it updates the topology graph in the system db with an
async job. In the window between we read the `lastClosedSystemTxId` and we read the `lastReconciledTxId( member )`
the TopologyGraphUpdater might write a new transaction to the system db which gets reconciled. In this case the
equality condition in the test assertion will never succeed. Changed to allow the `lastReconciledTxId( member )` to
be greater than or equal to.