GBPTree supports parallel writes
By the introduction of GBPTree#parallelWriter().
Multiple concurrent parallel writers may co-exist and write concurrently
to the tree, but not at the same time as the single-writer
GBPTree#writer(). Internally the coordination between parallel writers
is managed using a variant of latch crabbing technique with an added
optimistic mode which in the majority of cases uses shared latches
on parent nodes, for increased throughput.