MetaDataStore does not refresh fields at runtime
MetaDataStore has a cache of (almost) all fields and had this notion
of checking in all getters if the given field had been initialized,
and if not call "refreshFields" which would read all fields from store
and cache them. This was a silent problem because not all methods
that changed these fields wrote them to the store. I.e. reading
a field at runtime which wasn't initialized would reset all unstored
cached state to that of the stored state.

This is a dangerous behaviour since it could reset fields
seemingly at random, but for a long time it just so happened to not
be a problem because all fields were refreshed at the first "get"
of any field and would not be called after that.

With the introduction of database UUID and external store UUID
this changed a bit because:

- database UUID didn't have a field at all
- external store UUID didn't update the cached field on setting it,
  and could also be "not initialized" when upgrading a store from a
  pre-external-store-UUID era

The particular issue was this, a race between
nextCommittingTransactionId and getExternalStoreId()

- T1 calls nextCommittingTransactionId. lastCommittingTxIdField,
  an AtomicLong field governing this state) is e.g. 6 does
  incrementAndGet and returns 7.
- T2 performs some operation which accesses getExternalStoreId(),
  where the logic was that if the cached state of this field
  was "not initialized" it would call "refreshFields" and replace
  all cached state with that of the store. E.g. the field governing
  lastCommittingTxField would get reset to 6.
- T3 calls nextCommittingTransactionId and will also get back 7.
= There are now multiple transactions committing with transaction ID 7.

The fix is to:

- remove the dangerous notion of refreshing fields at runtime
- let all state have cached fields (database UUID didn't)
- let setting external store UUID also update the cached field
  (like all other setters).
- read all fields when opening the meta data store