Fixes an issue importing large duplicate IDs
Where if an ID from the e.g. CSV input file was large and required
a dynamic record (when stored in the temporary ID property store)
the wrong StoreCursors were passed in so that it would write into
the wrong store. This didn't cause other problems because any
property data for the node would be written afterwards and overwrite
it, but the node ID --> actual ID would throw exception later
if any of those large IDs had duplicates.