Change CoordinateReferenceSystem to enum
CoordinateReferenceSystem implements the CRS interface; and is only
accessed via the `public static final` instantiations within the class;
constructed with a `private` constructor.

The internals of the class static methods, are written explicitly to
assume that only those known instantiations exist.

This not only is conceptually an enum, it effectively already is; but
without Java being able to take full advantage of it.

Rather than changing everything to use Enum::name; I have kept the
behaviour the same to ensure logs and serialisation remains the same.