Don't add new slots for argument aliases in union
When the shared argument slots to both the left-hand side and right-hand side
of a Union contains variable name aliases, we do not need to create a new
separate slot for the alias.

This fixes a bug when a Union under an Apply with argument slots containing
aliases is followed by an Optional. When no row is produced (matched), Optional
manufactures a row with only argument values and null-values. In this case the
newly created slot for the variable name alias was only available when an
actual row was produced out of the Union. But with this fix it should now be
able to refer to the aliased argument variable in its original slot, which
should work also for the non-matched rows.

Co-Authored-By: Henrik Nyman <henrik.nyman@neotechnology.com>
Co-Authored-By: Alex Averbuch <alex.averbuch@gmail.com>
Co-Authored-By: Izabela Śmietana <izabela.smietana@neotechnology.com>