Page cursors payloads and logical offsets
Introduce the concept of payload to page cache pages.
Payload is the amount of space that can be used to store user content.
Payload is always smaller or equal to the page size.

Starting with this PR all of the page cache offsets are logical.
And all underlying page cursors do their part of the work in translating logical into real physical offsets.
All of the code that was using page sizes to evaluate offsets or the number of records per page was updated to use
the new payload concept.

Please note that this PR keeps the payload to be equal to page size.