Use default format selected by format selector instead of hardcoded
Before this PR a lot of tools and tests were using hardcoded default format or version strings that were not even correct one and
were testing things that are not even relevant for most users.
This PR switches those places to use dynamic neo4j format selector
that is used for default format selection for new stores.