Update EditionLocksFactories.java
The org.eclipse.jetty.util.StringUtil.isEmpty doesn't exists and cause the embedded usage to not work. The used class contains a isBlank method that checks if the string is empty or contains only whitespace characters. Since java 11, the java core String class contains a isBlank method that checks the same. So, I removed the StringUtils dependence and changed the isEmpty to java String default isBlank.