Caffeine-based LFU cache.
The previous implementation was using plain LinkedHashMap that was guarded by synchronized on any
access. That was not working that well in a concurrent environment where all the LRU caches that we have been used.
The updated implementation uses caffeine that is much more concurrency friendly.
Please note, that the new implementation of cache is not LRU, but LFU instead: https://arxiv.org/pdf/1512.00727.pdf