Database start context
Before this PR during database startup any party that required
cursorContext was creating one itself. Except participating in context creation anarchy
that has highlighted an issue with incorrect version context usage that was way too easy to use.
This PR start changing that by deprecating constructors that were supposed to be used only by tests
and introducing a new top-level component responsible for proper context creation and start using that across the stack.

The first set of contexts that were selected to begin clean-up with was database start.
Now the corresponding context propagated for all components that are started in a single thread as part of a particular db start.

However, there are known places where stub contexts are used in this PR to be able to split huge work into separate chunks.