Prevent *PartitionScanTestSuites running in Intellij
These test suites are intended to be run via PartitionScanTest, which
extent and construct with the correct arguments. TeamCity handles this
correctly.

It seems Intellij's test runner automatically detects the usage of jUnit
(most likely the use of @Nested) in these concrete classes, and treats
them as test files; despite having a constructor it doesn't know how to
deal with.

This is an issue if you try and run the "test" directly, or indirectly
via running all tests in a directory.

Making them abstract, despite technically being concrete, hides them.
This was overlooked when initially written; their parent classes were
already written as such despite otherwise also being concrete; for the
same reason.