Make neo4j-cypher-ir compile on Scala 2.13 ()
This is some of the changes from my test branch that make neo4j-cypher-ir compile. I don't know this code super-well so it might need a bit more work.

I've tried to avoid introducing copies - the only one is in QueryGraph.mutatingPatterns. This means less immutable collections because I've not messed with the ArrayBuffer pattern that is used in several places - I assume for a reason. This might not be what we want.