Add isAvailable() to GraphDatabaseService
Before this PR GraphDatabaseService had only `isAvailable(long timeoutMillis)`
method to check availability of the database, where timoutMillis was expected to be
availability await timeout.
In the real world, in the majority of places, `0` was passed as a timeout value.
This PR adds convenience method `isAvailable` that is equivalent `to isAvailable(0)` but do not force
end-users to provide any sort of timeouts.