Serial merge for lucene indexes
Our lucene indexes were configured to merge segments in separate threads.
During population that is probably not the best choice. Since population is
done as a background job it is important that it doesn't disturb normal operations.
Now population of our lucene indexes by default respects the
GraphDatabaseInternalSettings.index_population_workers setting by doing the merges in
the current thread. This should keep down CPU usage but will be slower.
It can be turned off with the introduced unsupported setting or index_population_workers
can be increased instead.