Fix flaky PartitionedSeekTest
Semantic changes to RandomSupport::nextLong along with an attempt to
account for them, led to a situation where from+1 == high; which
resulted in a / by 0.

        from in [0, high)
          to in [from+1, high)

ensures that there is an actual range; however, the max value for
from == high-1; then [from+1, high) => [high, high) is an "invalid range"
for RandomSupport::nextLong.

        from in [0, high)
          to in [from, high)

doesn't suffer from this, as the max value for from == high-1;
[from, high) => [high-1, high) => {high-1}

It is possible that to == from; however, this is valid as the range for
GBPTree::partitionedSeek is [from, to); however, an exact match is
indicated with from == to, due to the generic nature of GBPTree KEYs.