Consistency checker tweaks page cache mem setting for optimal perf
When running the `check-consistency` admin command, and admin commands
in general the page cache memory setting is read from neo4j.conf.
That setting is typically set quite high to get good performance
when running the neo4j dbms. However when doing a consistency check
the optimal performance is reached when there's a certain amount of
available off-heap memory for special caches manages by the checker.
The consistency checker can now spot if that setting is set too high
for this scenario and reduce it, resulting in better consistency check
performance.