Fix public API for Node::getRelationships so that its resource can be released
Node::getRelationships previously returned an Iterable, despite the primary
implementation managing a resource.

    Iterator<Relationship> rels = getRelationships().iterator();
    while ( rels.hasNext() )
    {
        // something that exits early;
        // a break, or an exception, etc.
    }

Though the resource would be closed when exhausted, it would not otherwise.
This would leak the underlying resource (ex. a cursor) and has been noticed.

Updating the return type to a ResourceIterable exposes the fact that any
iterators returned from these iterables must be closed. This ensures the
release of any underlying resources used by that iterator and hence prevent
any resource leaks.

    try ( ResourceIterator<Relationship> rels = getRelationships().iterator() )
    {
        while ( rels.hasNext() )
        {
            // something that exits early;
            // a break, or an exception, etc.
        }
        
    }  // now `rels` is correctly closed, no more leaking