KernelTransactions registers tx memory pool in init
Instead of in constructor. This fixes a problem where encountering
an error during Database#init would not remove that pool, i.e.
it would not properly clean up after itself. This would poison
the in-memory state of the dbms such that that particular database
could not be started again (at least if e.g. metrics were enabled
since there would now be two pools (the second registered on the
second start of that db) for which to register metrics,
and the metrics registry is global, requiring unique names.