Extend ResourceIterable such that it can release its resource
ResourceIterable returned a ResourceIterator that should be closed to prevent
any resource leaks. Unfortunately, this meant that the common Java idiom of
using for-each loops on an iterable could NOT be used as the underlying
resource would never be released.

This change exposes the Resource interface directly on the ResourceIterable so
that it may now be closed and for-each loops used safely. The implementtion
ensures that any iterators created by this iterable be closed when the parent
iterable is closed.

    try ( ResourceIterable<Node> nodes = getAllNodes() )
    {
        for (Node node : nodes )
        {
            // do something with nodes here
        }
        for (Node node : nodes )
        {
            // do something else with nodes here
        }
    }
    // both iterators created above will also be released now

This change has also refactored out the notion of monitoring of resources from
the tracking and closing of those resources. A ResourceMonitor will now follow
a Resource through the register/unregisterCloseableResource methods. A
ResourceTracker is a ResourceMonitor that can also release the monitored
resources through a call to closeAllCloseableResources. The API has been split
in this fashion to allow interested parties to monitor resources whilst also
keeping their lifecycle management internal.

Further to this, CursorIterator has been renamed to TrackedCursorIterator and
a base class introduced that performs no tracking of the cursor. This is for
the case where the lifecycle is managed by a wrapper component, such as a
ResourceIterable.