Non page cache based metadata store
Metadata store is specific store that carries around some values as key value pairs.
We had multiple problems with metadata store before this PR:
 * Metadata store fields are not covered by transaction log commands. As result recovery is a bit different for those fields or does not exist at all.
 * Metadata fields currently needs to be refreshed and that is a source of bugs
 * Metadata store is page cache backed, and we do not need those at all since that is sort of problematic now and will create problems for multi versioned store:
    - we need to open new non cached cursor for each interaction with this store and that is like that at least twice per transaction
    - we rewrite whole store (1 page) on flushes and any rewrite is a possibility to have junk in a fields that non-covered by recovery
    - almost any change in metadata store is another revision that will imply page version and extremely hot page that we want to avoid

This PR split metadata store into 2 parts:
- Part that is created for a new store (updated by clustering forced methods as well at this point) and never updated during the database lifetime.
That includes store id, database id, external id.
- Dynamic part that is fully inferred from transaction log tail and not stored anywhere else. Those values are the initial state of the world for metadata provider that is dynamically updated but are never stored since we always can restore them after any type of restart.

The execution model will change a bit as a result of this PR.
We will always start with looking up tail information (we do this now as well, but we ignore that in most of the scenarios outside of recovery or required recovery checks).
And will use that information to initialize metadata provider and other components. That information will be used as a single source of truth for dynamic values.
Cases of missing logs should be equivalent to the current way of handling missing logs and will mean that we can start with some initial values since the store is in an unknown state anyway. And some initial common state should make things more clear instead.