Rewrite any() and none() to IN where possible
Expressions like `any(x IN list WHERE x = 1)` and `none(x IN list WHERE x = 1)`,
where the predicate is a simple equality with the scope variable on one side
and something unrelated to the scope variable on the other, can be rewritten
into an `IN`. The runtime implementation of IN is more optimised, particularly
in pipelined runtime.