Buffered IDs for freeing can be offloaded to file
This caps the memory required for holding IDs that have been deleted,
but are not yet eligible to be reused, to tens or at most hundreds
of kilobytes instead of being unbounded. The previous unbounded nature
of this buffering could cause OOM when combining lots of deletions
and any transactions being open a long time.