Unflake RecoveryCorruptedTransactionLogIT
The test recoverFromLastCorruptedBrokenCheckpointRecord removed
the last checkpoint and replaced it with random bytes.
Sometimes those random bytes would be interpreted as a proper
checkpoint, or "not a checkpoint" since it was fine until
hitting end-of-file and therefore not be seen as corrupted.
The test asserts that startup sees a corrupted checkpoint,
so therefore the test now verifies that the random bytes looks
corrupted before writing those garbage bytes and starting up.