Optimise caching of IN rhs
In some cases we build up a set of the rhs in IN expressions. In many cases
this is expensive without any benefits. But there are queries where it's useful,
one example could be `MATCH (a:A)-->(b:B) WHERE b.prop IN a.list RETURN a;`.
This is an attempt to only use caching when it's useful.

- Skip caching on small lists.
- Use reference identity as cache key to avoid having to compare all values in
  list to get a cache hit.
- Start expensive set buildup on second call.