At least 100x faster relationship group degrees store rebuild
If relationship group degrees store is missing it needs to be
rebuilt during startup. Previously this rebuild was implemented
in a naive way which made it very slow. By using techniques
like the counts store rebuild and algorithms that runs through
the stores sequentially this is now several orders of magnitudes
faster.