DEFAULT should be HOME in SHOW PRIVILEGES ()
* DEFAULT should be HOME in SHOW PRIVILEGES

Priviliges against the HOME psudodatabase should not be shown as linked to DEFAULT. This is due to the `name` property on the `:DatabaseDefault` node being set to 'DEFAULT'. This change removes that property from both `:DatabaseAll` and `:DatabaseDefault` nodes and works out the database type by looking at the labels.