Allow metadata store format to forcefully read records
Before this PR during migration fields that have a position that is higher than the current position were ignored/read as non-used fields. This PR introduces the possibility to read records in FORCE mode to still read those records to be able to get values of legacy fields during migration.