Disable `allTimeZonesAreValidZoneIDs` test
This test is too restrictive as-is: Zone IDs aren't stable across JDKs, 'Pacific/Kanton' isn't currently supported by x86-ubuntu-oraclejdk-17. I think it points to an underlying issue with how we handle timezones and that we should enable this test, or some variation of it, once we have a better solution.