Don't do 4.x->5.x migration when migrating between format families on 5.x
When migrating for example from standard 5.0 -> aligned 5.0 we don't need to do the 4.x -> 5.x migration. We would fail in trying to do this because we would read kernel version from 5.0 metadata store as if we were reading from 4.x metadata store, using the old ids, and we would fail getting kernel version from -1.