Use kernel version from old store when migrating indexes
Before, the metadata store which is newly created on 5.0 migration, would report kernel version as 5.0. But the old schema store that we are migrating should be read through the lens of the kernel version on the old store, so we read this kernel version and set it temporarily during migration of indexes.