Upgrade of the database with legacy transaction logs
Enhance upgrade to be able to upgrade the database with transaction logs
that have some commands written in formats that are not supported anymore.
To be able to upgrade such databases we need the last transaction to be written in a supported format.
We will try to work around the inability to read the whole log file by reversing the log file and attempting to parse the last commit entry.