Mark multi versioned store prototype affected tests
Some tests are supposed to fail at the moment with a multi versioned store prototype.
To be able to have a build that works on that prototype we introduce a new tag to mark all tests like that.
Those marked tests will be excluded from the set of tests that are executed in that specific branch.