Fix literal string comparison bug () ()
There was a bug which existed here where the IndexHints would return an exception. The bug was because the source code would compare strings without ignoring cases. This solution fixes the problem. A small analysis was done to investigate whether a wider refactor was needed, and it found that a larger refactor was not needed (https://github.com/neo-technology/daniel-notepad/blob/main/literal-string-comparison-issue/analysis.md)

There was a conflict when merging because of missing tests in IndexPlanningIntegrationTest.scala. The tests were missing because the underlying behaviour they test is only present in 5.0

Cherry-picks: