Fix AST existential subclause pattern position ()
The ASTParser would not assign the correct position to the pattern within an existential subquery. We tried to fix the problem in the JavaCC directly but it was not possible to do so on account of the pattern expression being able to take multiple forms. In the end this was the best approach.

(cherry picked from commit 2c983df4b498c10f08fbac9ea9344d18619bf9e7)