Extend ResourceIterable such that it can release its resource
ResourceIterable returned a ResourceIterator that should be closed to prevent
any resource leaks. Unfortunately, this meant that the common Java idiom of
using for-each loops on an iterable could NOT be used as the underlying
resource would never be released.

This change exposes the Resource interface directly on the ResourceIterable so
that it may now be closed and for-each loops used safely. The implementation
ensures that any iterators created by this iterable be closed when the parent
iterable is closed.

    try ( ResourceIterable<Node> nodes = getAllNodes() )
    {
        for (Node node : nodes )
        {
            // do something with nodes here
        }
        for (Node node : nodes )
        {
            // do something else with nodes here
        }
    }
    // both iterators created above will also have been released now

A base ResourceIterable class has been also introduced that tracks any
iterators produced and closes them when the iterable itself is closed. This is
useful when employing a for-each loop to iterate over an iterable to ensure
that the iterator created as part of that operation is also closed.

Tracking of ResourceIterables created in Node via the resource monitor of a
transaction was also investigated (allowing all iterable resources to be
released when the transaction is closed) but this had a noticable performance
impact in some benchamrks so is not currently enabled. A future PR will
investigate reintroducing this behaviour but controlled via some system flag.

The Iterables and Iterators utility classes have also been updated to ensure
that (when appropriate) any Iterables/Iterators that implement Resource are
closed after they have been processed in the utilities methods.

Further to this, CursorIterator has been renamed to TrackedCursorIterator and
a base class introduced that performs no tracking of the cursor. This is for
the case where the lifecycle is managed by a wrapper component, such as a
ResourceIterable.

This commit also includes the requisite changes to the calls site usage of
ResourceIterables by closing them. This includes calls to the various
getRelationships(Direction) methods in Node and getAllNodes and
getAllRelationships in Transaction.

NOTE This change unreverts a previous commit that updates ResourceIterable to
also be a Resource () . That previous commit had severe regressions in
that the many usages of ResourceIterable were NOT also being closed leading to
severe GC issues in many benchmarks.