Implement following partition scan for token indexes
Given a PartitionScan for a token index, create another that is
partitioned in a way such that each entity containing the queried token,
will be uniquely referenced in the corresponding partition.

This is implemented in a way that the returned PartitionScan (follower)
is dependent on the partitioning of the given PartitionScan (leader).
To ensure the best usage of resources, the given PartitionScan should
correspond to the query with the most expected data.

Currently the implementation is that the follower has the exact same
partitioning ranges as the leader.

Also added a helper method that will create a list of PartitionScans,
corresponding to the array of TokenPredicates given.

Co-authored-by: Anton Persson <anton.persson@neotechnology.com>