Relationship type expressions
Introduced colon disjunction for legacy relationship type disjunctions '|:'

LabelExpressions are no Expressions any more but rather normal ASTNodes

A rewriter rewrites relationship types which do not conform to previous cypher syntax into predicates. Those relationship types only consisting of disjunctions are turned into relationship type sequences as before.

Mixing the different syntaxes is considered an error. We can change that if the CIP process demands a change.

Relationship type expressions are not allowed on variable length relationships for now.

Implemented an algorithm to determine whether two label expressions overlap in AddUniquenessPredicates.

Stringifier inserts parentheses if and only if operators are different. We use parentheses to express precedence of | and &.

Report a legible error message when using wildcard in label expression predicate

Disallow relationship type expressions in shortest path
We need some type on the label expression leaves that is different from symbolic names to not rewrite indefinitely on the label name in the normalizer

Co-authored-by: Bastien Louërat <bastien.louerat@neo4j.com>