Counts store doesn't need to be rebuilt after import
At the end of an import the counts store would be built.
The problem was that the "last transaction id" it was marked
with was wrong, because right afterwards there would be appended
an empty "upgrade" transaction. Therefore on the next start of
the db the counts store would need to be rebuilt again.