Simplify flaky import test
Remove dependency on parameter evaluation order to make the test simpler
to read.

Remove reset of overriden input ids in ControlledEncder class used for
testing. This test class is only used in the flaky test and that test
always sets the values.

This will probably not fix the flakyness of the test unless the problem
lies somewhere in resetting the overriden id in Controlled encoder.