GBPTree supports multiple roots
The design is that there's a "root layer" which manages access to
data trees. One data tree is like the typical write/seek access
to a GBPTree.

Each created data tree is completely isolated from the other trees
and are created, updated and deleted explicitly by the user
and all data trees live in the same GBPTree file, sharing the freelist.

Keys for external roots in multi-root GBPTree are customizable using
the a separate layout as well.

GBPTree still exists just like before, but has a super class
MultiRootGBPTree which is more generic, where GBPTree is a thin subclass
narrowing the functionality down to a single root

Additionally moved read-only check for writer out of GBPTree
so that there doesn't need to a concept of "unsafe" writer.
Only index accessors currently need to check read-only state,
that's fewer places than used the "unsafe" writer.