Relationship type index/scan can be based around node IDs instead
Such that each "bit" in the relationship type lookup index is a node ID
and the token ID is which relationship type that node has.
Kernel scan cursors in this area have been changed to:

 - Look and feel more natural w/ regards to accessing relationship data
   for each entry, instead of revolving around IDs and forcing the
   read via an external RelationshipScanCursor
 - Have variants that traverses nodes to get relationships for the
   given type for each entry in the lookup index

Both strategies works regardless of storage engine and ultimately
it's up to the storage engine to tell which strategy is used and
to generate the appropriate lookup index updates accordingly.