"Inner" cursors aren't pooled
Pooling - let's consider the simple case of, say a simple
DefaultNodeCursor. It has no internal "kernel" cursor, only
the usual storage cursor(s). When this node cursor is closed
it's returned to the pool for others to use.

Now consider a more complex cursor DefaultRelationshipTypeIndexCursor
which has other internal kernel cursor. Those internal cursors
have previously been instantiated without getting the "acquire"
call to them, but have still been returned to the pool when
the parent DefaultRelationshipTypeIndexCursor was released.
Also having internal cursors mixed with "external/actual" cursors
in the pool makes the pooling unnecessarily complicated to think about.

This changes so that it's only the "external/actual" cursor,
the one that you're allocating that is pooled and its internal
cursors only live within it, not in the pool. This makes things
simpler and also avoids unnecessary churn of basic node/relationship
cursors.