Replace DatabaseManager with DatabaseContextProvider ()
* Replace DatabaseManager with DatabaseContextProvider

We don't want to pass around and API for managing databases, it should only be consumed by the reconciler. Today the DatabaseManager is used in many places, but only for looking up databases.
This is solved by extracting the provider interface from the DatabaseManager. The provider interface is called DatabaseContextProvider. The DatabaseManger is only used by the reconciler now and not added to dependencies.

Simplify the DatabaseManager implementations. There is a lot of inheritance here with unnecessary complicated code for no reason. The current two implementations of the database manager are doing completely different things and should not share an interface.
This is done by removing the DatabaseManger interface and the AbtractDatabaseManager. DatabaseManager is now just a class only used in enterprise and cluster where multi database is allowed. DatabaseManger is only used by the reconciler now and not added to dependencies.

Community version starts and stops databases through lifecycle and does not need a DatabaseManager at all. This is done through DatabaseLifecycles class.