Removes virtually unused accidental complexity from progress listener
It had a concept of "started" which was unnecessary to expose
to the user of the ProgressListener.

There was also a concept of "started" w/ a name, which was added
for an over-engineered solution of progress reporting of a
certain part of the batch importer, which is now not visible
to the user in this way (with the name changes mid-progress) anyway.